requirements.txt
straight.plugin==1.4.0-post-1
Biopython
helperlibs
nose
